requirements.txt
fastapi==0.127.1
uvicorn==0.30.6
openai==2.13.0
pydantic==2.7.4
python-dotenv==1.0.1
